requirements.txt
matplotlib
python-mnist
scipy
torch
torchvision
tqdm